feat: move optional bindings out of rollup and list as exports
