+++ packages/web-components/fast-element/docs/api-report.md
@@ -295,12 +295,6 @@ export type DecoratorAttributeConfiguration = Omit<AttributeConfiguration, "prop
 export type DefaultBindingOptions = AddEventListenerOptions;
 
 // @public
-export type DefaultTwoWayBindingOptions = DefaultBindingOptions & {
-    changeEvent?: string;
-    fromView?: (value: any) => any;
-};
-
-// @public
 export interface Disposable {
     dispose(): void;
 }
@@ -711,16 +705,6 @@ export interface RootContext {
 }
 
 // @public
-export const signal: <T = any>(options: string | Binding<T, any, ExecutionContext<any>>) => BindingConfig<string | Binding<T, any, ExecutionContext<any>>>;
-
-// @public
-export class SignalBinding extends UpdateBinding {
-    bind(source: any, context: ExecutionContext, targets: ViewBehaviorTargets): void;
-    static send(signal: string): void;
-    unbind(source: any, context: ExecutionContext, targets: ViewBehaviorTargets): void;
-}
-
-// @public
 export function slotted<T = any>(propertyOrOptions: (keyof T & string) | SlottedDirectiveOptions<keyof T & string>): CaptureType<T>;
 
 // @public
@@ -852,23 +836,6 @@ export type TrustedTypes = {
 export type TrustedTypesPolicy = {
     createHTML(html: string): string;
 };
-
-// @public
-export const twoWay: BindingConfig<DefaultTwoWayBindingOptions> & BindingConfigResolver<DefaultTwoWayBindingOptions>;
-
-// @public
-export class TwoWayBinding extends ChangeBinding {
-    bind(source: any, context: ExecutionContext, targets: ViewBehaviorTargets): void;
-    static configure(settings: TwoWaySettings): void;
-    // @internal (undocumented)
-    handleEvent(event: Event): void;
-    unbind(source: any, context: ExecutionContext, targets: ViewBehaviorTargets): void;
-}
-
-// @public
-export interface TwoWaySettings {
-    determineChangeEvent(directive: HTMLBindingDirective, target: HTMLElement): string;
-}
 
 // Warning: (ae-internal-missing-underscore) The name "TypeDefinition" should be prefixed with an underscore because the declaration is marked as @internal
 //
