+++ packages/web-components/fast-element/src/templating/binding.ts
@@ -1,10 +1,9 @@
-import { isString, Message } from "../interfaces.js";
+import { Message } from "../interfaces.js";
 import {
     Binding,
     BindingObserver,
     ExecutionContext,
     Observable,
-    ObservationRecord,
 } from "../observation/observable.js";
 import { FAST } from "../platform.js";
 import { DOM } from "./dom.js";
@@ -366,89 +365,6 @@ export class OneTimeBinding extends UpdateBinding {
     }
 }
 
-const signals: Record<string, undefined | Function | Function[]> = Object.create(null);
-
-/**
- * A binding behavior for signal bindings.
- * @public
- */
-export class SignalBinding extends UpdateBinding {
-    private handlerProperty = `${this.directive.id}-h`;
-
-    /**
-     * Bind this behavior to the source.
-     * @param source - The source to bind to.
-     * @param context - The execution context that the binding is operating within.
-     * @param targets - The targets that behaviors in a view can attach to.
-     */
-    bind(source: any, context: ExecutionContext, targets: ViewBehaviorTargets): void {
-        const directive = this.directive;
-        const target = targets[directive.nodeId];
-        const signal = this.getSignal(source, context);
-        const handler = (target[this.handlerProperty] = () => {
-            this.updateTarget(
-                target,
-                directive.targetAspect!,
-                directive.binding(source, context),
-                source,
-                context
-            );
-        });
-
-        handler();
-
-        const found = signals[signal];
-
-        if (found) {
-            Array.isArray(found)
-                ? found.push(handler)
-                : (signals[signal] = [found, handler]);
-        } else {
-            signals[signal] = handler;
-        }
-    }
-
-    /**
-     * Unbinds this behavior from the source.
-     * @param source - The source to unbind from.
-     * @param context - The execution context that the binding is operating within.
-     * @param targets - The targets that behaviors in a view can attach to.
-     */
-    unbind(source: any, context: ExecutionContext, targets: ViewBehaviorTargets): void {
-        const signal = this.getSignal(source, context);
-        const found = signals[signal];
-
-        if (found && Array.isArray(found)) {
-            const directive = this.directive;
-            const target = targets[directive.nodeId];
-            const handler = target[this.handlerProperty];
-            const index = found.indexOf(handler);
-            if (index !== -1) {
-                found.splice(index, 1);
-            }
-        } else {
-            signals[signal] = void 0;
-        }
-    }
-
-    private getSignal(source: any, context: ExecutionContext): string {
-        const options = this.directive.options;
-        return isString(options) ? options : options(source, context);
-    }
-
-    /**
-     * Sends the specified signal to signaled bindings.
-     * @param signal - The signal to send.
-     * @public
-     */
-    public static send(signal: string): void {
-        const found = signals[signal];
-        if (found) {
-            Array.isArray(found) ? found.forEach(x => x()) : found();
-        }
-    }
-}
-
 /**
  * A binding behavior for bindings that change.
  * @public
@@ -610,103 +526,6 @@ export class EventBinding {
 }
 
 /**
- * The settings required to enable two-way binding.
- * @public
- */
-export interface TwoWaySettings {
-    /**
-     * Determines which event to listen to, to detect changes in the view.
-     * @param directive - The directive to determine the change event for.
-     * @param target - The target element to determine the change event for.
-     */
-    determineChangeEvent(directive: HTMLBindingDirective, target: HTMLElement): string;
-}
-
-let twoWaySettings: TwoWaySettings = {
-    determineChangeEvent() {
-        return "change";
-    },
-};
-
-/**
- * A binding behavior for bindings that update in two directions.
- * @public
- */
-export class TwoWayBinding extends ChangeBinding {
-    private changeEvent: string;
-
-    /**
-     * Bind this behavior to the source.
-     * @param source - The source to bind to.
-     * @param context - The execution context that the binding is operating within.
-     * @param targets - The targets that behaviors in a view can attach to.
-     */
-    bind(source: any, context: ExecutionContext, targets: ViewBehaviorTargets): void {
-        super.bind(source, context, targets);
-
-        const directive = this.directive;
-        const target = targets[directive.nodeId] as HTMLElement;
-
-        if (!this.changeEvent) {
-            this.changeEvent =
-                directive.options.changeEvent ??
-                twoWaySettings.determineChangeEvent(directive, target);
-        }
-
-        target.addEventListener(this.changeEvent, this);
-    }
-
-    /**
-     * Unbinds this behavior from the source.
-     * @param source - The source to unbind from.
-     * @param context - The execution context that the binding is operating within.
-     * @param targets - The targets that behaviors in a view can attach to.
-     */
-    unbind(source: any, context: ExecutionContext, targets: ViewBehaviorTargets): void {
-        super.unbind(source, context, targets);
-        (targets[this.directive.nodeId] as HTMLElement).removeEventListener(
-            this.changeEvent,
-            this
-        );
-    }
-
-    /** @internal */
-    public handleEvent(event: Event): void {
-        const directive = this.directive;
-        const target = event.currentTarget as HTMLElement;
-
-        let value;
-
-        switch (directive.aspectType) {
-            case 1:
-                value = target.getAttribute(directive.targetAspect);
-                break;
-            case 2:
-                value = target.hasAttribute(directive.targetAspect);
-                break;
-            case 4:
-                value = target.innerText;
-                break;
-            default:
-                value = target[directive.targetAspect];
-                break;
-        }
-
-        const observer = this.getObserver(target);
-        const last = (observer as any).last as ObservationRecord; // using internal API!!!
-        last.propertySource[last.propertyName] = directive.options.fromView(value);
-    }
-
-    /**
-     * Configures two-way binding.
-     * @param settings - The settings to use for the two-way binding system.
-     */
-    public static configure(settings: TwoWaySettings) {
-        twoWaySettings = settings;
-    }
-}
-
-/**
  * The default binding options.
  * @public
  */
@@ -722,43 +541,12 @@ export const onChange = BindingConfig.define(
 );
 
 /**
- * The default twoWay binding options.
- * @public
- */
-export type DefaultTwoWayBindingOptions = DefaultBindingOptions & {
-    changeEvent?: string;
-    fromView?: (value: any) => any;
-};
-
-/**
- * The default twoWay binding configuration.
- * @public
- */
-export const twoWay = BindingConfig.define(BindingMode.define(TwoWayBinding), {
-    fromView: v => v,
-} as DefaultTwoWayBindingOptions);
-
-/**
  * The default onTime binding configuration.
  * @public
  */
 export const oneTime = BindingConfig.define(BindingMode.define(OneTimeBinding), {
     once: true,
 } as DefaultBindingOptions);
-
-const signalMode: BindingMode = BindingMode.define(SignalBinding);
-
-/**
- * Creates a signal binding configuration with the supplied options.
- * @param options - The signal name or a binding to use to retrieve the signal name.
- * @returns A binding configuration.
- * @public
- */
-export const signal = <T = any>(
-    options: string | Binding<T>
-): BindingConfig<string | Binding<T>> => {
-    return { mode: signalMode, options };
-};
 
 /**
  * A directive that applies bindings.
